ghojeong(고정완) [Step4] 로또(수동) (#914)
* [fix] input 을 제대로 못 입력받는 문제 디버그

* [refactor]: LottoNoPool 과 Rank 미리 캐싱하도록 변경

* [feat]: 수동으로 추첨 번호를 입력하는 기능 구현

* [style]: 계산하고자 하는 두 값 중 하나만 double형으로 형변환

* [refactor]: PurchaseInfo 삭제

* [refactor]: WinningCondition.java 삭제

* [style]: 불필요한 함수 삭제

* [style]: Rename variable

* Revert "[style]: Rename variable"

This reverts commit 6fbf1b7aa7c47a9fc00f09970b92ec28ca819d95.

* Revert "[style]: 불필요한 함수 삭제"

This reverts commit c82fc7fbc8f72cb32f2e1f515d57722379b1dfee.

* Revert "[refactor]: WinningCondition.java 삭제"

This reverts commit 6d4a09fcb509e7d6237f9e562463f5312f5bd4b2.

* Revert "[refactor]: PurchaseInfo 삭제"

This reverts commit dd7f1b0d7b02e89321c12c756f3e24f51d3bf321.

* [refactor]: Rename PurchaseInfo to Receipt

+++ src/test/java/lotto/domain/MoneyTest.java
@@ -29,7 +29,7 @@ class MoneyTest {
         Money moneyObj = new Money(money);
         Money dividerObj = new Money(divider);
         assertThat(moneyObj.divide(dividerObj))
-                .isEqualTo(money / divider);
+                .isEqualTo((double) money / divider);
     }
 
     @Test
